Renames all hooks that don't start with use
