--- src/hooks/accounts/index.ts
@@ -1,4 +1,0 @@
-import accounts from './accounts'
-
-export * from './types'
-export default accounts
