--- src/hooks/network/index.ts
@@ -1,4 +1,0 @@
-import network from './network'
-
-export default network
-export * from './types'
