+++ src/hooks/useAccounts/index.ts
@@ -1,0 +1,4 @@
+import useAccounts from './useAccounts'
+
+export * from './types'
+export default useAccounts
