+++ src/hooks/useNetwork/index.ts
@@ -1,0 +1,4 @@
+import useNetwork from './useNetwork'
+
+export default useNetwork
+export * from './types'
